LayoutTests/platform/mac-snowleopard/plugins/embed-prefers-plugins-for-images-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: src: image.png
PASS
